requirements.txt
# Telegram Bot Framework
aiogram>=3.0.0

# HTTP Clients
aiohttp>=3.8.0
httpx>=0.24.0
requests>=2.31.0
requests-cache>=1.0.0
retry-requests>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# LLM and AI
openai>=1.0.0

# Audio Processing
faster-whisper>=0.9.0
flask>=3.0.0

# Document Processing and RAG
llama-index>=0.9.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-ollama>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Weather API
openmeteo-requests>=1.0.0

# Utilities
ipython>=8.12.0
